solved issue #32
